semantic_search/requirements.txt
sentence-transformers>=2.7,<3.0
faiss-cpu>=1.7
numpy>=1.26
python-dotenv>=1.0
